Arreglada gestión errores mensajes

+++ services/autenticacion.service.js
@@ -60,9 +60,10 @@ const service = {
 
       return { ...userWithClaims, ...userDB.data() };
     } catch (err) {
-      await admin.auth().deleteUser(user.uid);
-      console.error(err);
-      throw new Error(`Error al crear el usuario: ${err}`);
+      if (user && user.uid) await admin.auth().deleteUser(user.uid);
+
+      if (err && err.message) throw new Error(err.message);
+      else throw new Error(err);
     }
   },
 };
